# deliverables/openapi.yaml
openapi: 3.0.3
info:
  title: 墨阅小说网 API
  version: 1.0.0
  description: |
    免费读书网站「墨阅小说网」后端接口契约（OpenAPI 3.0）。
    前端 Vue 3 三端（读者端 / 作家端 / 管理后台）与本契约并行开发。
    约定：统一响应体 R<T>；鉴权 Authorization: Bearer {accessToken}；
    分页参数 page / size；写操作需带 Idempotency-Key；错误码为 5 位业务码。
servers:
  - url: https://api.moyue.com/api/v1
    description: 生产环境
  - url: https://staging-api.moyue.com/api/v1
    description: 预发环境
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: 手机号 + 密码登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginReq" }
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
        "401":
          description: 账号或密码错误
  /auth/register:
    post:
      tags: [auth]
      summary: 注册账号并下发短信验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code, password]
              properties:
                phone:    { type: string }
                code:     { type: string }
                password: { type: string, format: password }
      responses:
        "200": { description: 注册成功 }
...
  /chapters/{chapterId}:
    get:
      tags: [chapter]
      summary: 章节正文（热点缓存，未命中回源 MySQL 从库）
      parameters:
        - name: chapterId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: OK }
...
  /author/books/{bookId}/chapters:
    post:
      tags: [chapter]
      summary: 新建章节（默认草稿态，支持自动保存）
      parameters:
        - name: bookId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: 创建成功 }
...
  /comments:
    get:
      tags: [comment]
      summary: 评论列表（按书籍 / 章节分页，含点赞数）
      parameters:
        - name: bookId
          in: query
          schema: { type: integer, format: int64 }
        - name: chapterId
          in: query
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: OK }
    post:
      tags: [comment]
      summary: 发表评论，先落库待审再异步机审
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId, content]
              properties:
                bookId:    { type: integer, format: int64 }
                chapterId: { type: integer, format: int64 }
                content:   { type: string, maxLength: 1000 }
      responses:
        "200": { description: 已提交，审核中 }
...
  /admin/audit/comments/{id}:
    post:
      tags: [admin]
      summary: 审核通过 / 驳回，写回审核结果与通知
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      description: 统一响应体 R<T>
      properties:
        code:    { type: integer, example: 0 }
        message: { type: string,  example: success }
        data:    { type: object }
        traceId: { type: string }
    LoginReq:
      type: object
      required: [phone, password]
      properties:
        phone:    { type: string, pattern: "^1[3-9][0-9]{9}$" }
        password: { type: string, format: password, minLength: 6 }
...